we can set message variable w/ burnettk cullerton
